+++ repo/src/main/java/com/moonwatch/repo/AlertRepo.kt
@@ -5,6 +5,7 @@ import com.moonwatch.core.repo.IAlertRepo
 import com.moonwatch.db.dao.AlertDao
 import com.moonwatch.db.entity.TokenAlertEntity
 import dagger.Reusable
+import java.math.BigDecimal
 import java.util.*
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
@@ -16,8 +17,8 @@ class AlertRepo @Inject constructor(private val dao: AlertDao) : IAlertRepo {
 
   override suspend fun addAlert(
       address: String,
-      sellPriceTargetUsd: Double?,
-      buyPriceTargetUsd: Double?
+      sellPriceTargetUsd: BigDecimal?,
+      buyPriceTargetUsd: BigDecimal?
   ) {
     dao.insertAlert(
         TokenAlertEntity(
